security: collected low-severity fixes from review pass 2
* logger hook: bump message_length type from u8 to size_t so a log
  line longer than 255 bytes doesn't silently truncate the length
  passed to the hook
* is_safe_to_log: drop the bogus `bool inline` (no static linkage)
  -- the header already declared it as a normal function so the
  inline keyword was a no-op anyway
* mongo_authenticate: tolerate non-UTF8 entries in
  saslSupportedMechs by `continue`-ing instead of breaking the loop
  -- future server might add an entry of an unrelated type before
  a SCRAM mech we do recognize
* mongo_client_free: zeroize the full mongo_client_t (which holds
  a copy of username + password) before vPortFree, via
  mbedtls_platform_zeroize so the optimizer can't elide it
* CMake: wire MONGO_SCRAM_DEBUG as a proper option(...) so the
  -DMONGO_SCRAM_DEBUG=1 the docs advertise actually works

+++ src/app/logger_hook.c
@@ -9,13 +9,14 @@
  * accessible to a host tool, etc.) without touching call sites.
  */
 
+#include <stddef.h>
 #include <stdio.h>
 
 #include "logging_api.h"
 #include "logging_config.h"
 #include "types.h"
 
-void acw_post_logging_hook(char *message, u8 message_length) {
+void acw_post_logging_hook(char *message, size_t message_length) {
     (void)message_length;
 
 #if LOGGING_LOG_VIA_PRINTF
